+++ tasks/seq/sadikov_I_SparseMatrixMultiplication/func_tests/main.cpp
@@ -5,8 +5,8 @@
 #include <vector>
 
 #include "core/task/include/task.hpp"
-#include "seq/sadikov_I_SparesMatrixMultiplication/include/SparesMatrix.hpp"
-#include "seq/sadikov_I_SparesMatrixMultiplication/include/ops_seq.hpp"
+#include "seq/sadikov_I_SparseMatrixMultiplication/include/SparesMatrix.hpp"
+#include "seq/sadikov_I_SparseMatrixMultiplication/include/ops_seq.hpp"
 
 TEST(sadikov_i_sparse_matrix_multiplication_task_seq, test_rect_matrixes) {
   constexpr auto kEpsilon = 0.000001;
