+++ tasks/seq/sadikov_I_SparseMatrixMultiplication/include/SparesMatrix.hpp
@@ -21,7 +21,7 @@ class SparesMatrix {
         m_columnsCount_(columns_count),
         m_values_(values),
         m_rows_(rows),
-        m_elementsSum_(element_sum) {};
+        m_elementsSum_(element_sum){};
   [[nodiscard]] const std::vector<double>& GetValues() const noexcept { return m_values_; }
   [[nodiscard]] const std::vector<int>& GetRows() const noexcept { return m_rows_; }
   [[nodiscard]] const std::vector<int>& GetElementsSum() const noexcept { return m_elementsSum_; }
