+++ tasks/seq/sadikov_I_SparseMatrixMultiplication/include/ops_seq.hpp
@@ -5,16 +5,15 @@
 #include <vector>
 
 #include "core/task/include/task.hpp"
-#include "seq/sadikov_I_SparesMatrixMultiplication/include/SparesMatrix.hpp"
+#include "seq/sadikov_I_SparseMatrixMultiplication/include/SparesMatrix.hpp"
 
 namespace sadikov_i_sparse_matrix_multiplication_task_seq {
-
 std::vector<double> GetRandomMatrix(int size);
 class CCSMatrixSequential : public ppc::core::Task {
   SparesMatrix m_fMatrix_;
   SparesMatrix m_sMatrix_;
   SparesMatrix m_answerMatrix_;
-  // restart tests
+
  public:
   explicit CCSMatrixSequential(ppc::core::TaskDataPtr task_data) : Task(std::move(task_data)) {}
   bool PreProcessingImpl() override;
